refactored: new tooltip.js
